implementing some helpers for: info, info commandstats

+++ src/System/DevUtils/Redis/Helpers/CommandStats/Include.hs
@@ -5,6 +5,6 @@ module System.DevUtils.Redis.Helpers.CommandStats.Include (
 data CommandStat = CommandStat {
  _type :: String,
  _numCalls :: Integer,
- _totalCpu :: Float,
- _avgCpu :: Float
+ _totalCpu :: Integer,
+ _avgCpu :: Double
 } deriving (Show)
